% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARD_script_example.R
\name{ARD_script_example}
\alias{ARD_script_example}
\title{Get path to example ARD scripts, auto-generated by siera}
\usage{
ARD_script_example(path = NULL)
}
\arguments{
\item{path}{Name of R file. If `NULL`, the example files will be listed.}
}
\value{
A list of example R files (if path is NULL), or a file itself if path is used.
}
\description{
siera comes bundled with some example ARD-producing scripts in
its `inst/script` directory. This function make them easy to access.
}
\examples{
ARD_script_example()
ARD_script_example("ARD_Out14-1-1.R")
}
