% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysisMethods.R
\name{.generate_analysis_method_section}
\alias{.generate_analysis_method_section}
\title{Format numeric values consistently}
\usage{
.generate_analysis_method_section(
  analysis_methods,
  analysis_method_code_template,
  analysis_method_code_parameters,
  method_id,
  analysis_id,
  output_id,
  envir = parent.frame()
)
}
\arguments{
\item{analysis_methods}{AnalysisMethod dataset for the reporting event}

\item{analysis_method_code_template}{AnalysisMethodCodeTemplate dataset for the reporting event}

\item{analysis_method_code_parameters}{AnalysisMethodCodeParameters dataset for the reporting event}

\item{method_id}{MethodId for the method applied to current Analysis}

\item{analysis_id}{AnalysisId for current Analysis}

\item{output_id}{OutputId to which current Analysis belongs}

\item{envir}{Environment (parent environment as default)}
}
\value{
Character vector with formatted numbers.
}
\description{
Internal helper to apply consistent number formatting across analyses.
}
\keyword{internal}
