% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readARS.R
\name{readARS}
\alias{readARS}
\title{readARS is a function that ingests ARS metadata, and meta-programmes R code that could be run as-is to produce Analysis Results Datasets.}
\usage{
readARS(JSON_ARS, output_path = "")
}
\arguments{
\item{JSON_ARS}{A JSON file containing ARS metadata for a reporting event.}

\item{output_path}{Path to store .R ARD scripts}
}
\value{
R programmes generating ARDs - one for each output specificied in the ARS JSON
}
\description{
readARS is a function that ingests ARS metadata, and meta-programmes R code that could be run as-is to produce Analysis Results Datasets.
}
\examples{
# get path to json file
json_path <- system.file("extdata", "ARS_V1_Common_Safety_Displays.json", package = "siera")

# output path for R programs
output_dir = file.path(tempdir())

# run function, write to temp directory
readARS(json_path, output_dir)

#remove temp directory
unlink(output_dir, recursive = TRUE)

}
