% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readARS.R
\name{readARS}
\alias{readARS}
\title{readARS}
\usage{
readARS(JSON_ARS)
}
\arguments{
\item{JSON_ARS}{- A JSON file containing ARS metadata for a reporting event.}
}
\value{
- R programmes generating ARDs - one for each output specificied in the ARS JSON
}
\description{
readARS
}
\examples{
- readARS("ARS_V1_Common Safety Displays.json")

}
