% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readARS.R
\name{readARS}
\alias{readARS}
\title{readARS is a function that ingests ARS metadata, and meta-programmes R code that could be run as-is to produce Analysis Results Datasets.}
\usage{
readARS(JSON_ARS)
}
\arguments{
\item{JSON_ARS}{A JSON file containing ARS metadata for a reporting event.}
}
\value{
R programmes generating ARDs - one for each output specificied in the ARS JSON
}
\description{
readARS is a function that ingests ARS metadata, and meta-programmes R code that could be run as-is to produce Analysis Results Datasets.
}
\examples{
json_path <- system.file("extdata", "ARS_V1_Common_Safety_Displays.json", package = "siera")
readARS(json_path)

}
